update admin interface

+++ src/main/webapp/ManageListVaccine.jsp
@@ -14,11 +14,7 @@
         <link rel="shortcut icon" type="image/png" href="../assets/images/logos/favicon.png" />
         <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.3/css/all.min.css" />
         <link rel="stylesheet" href="./assets/css/AdminIndex.min.css" />
-        <script>
-            $(document).ready(function () {
-                $('[data-toggle="tooltip"]').tooltip();
-            });
-        </script>
+        <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/sweetalert2@11/dist/sweetalert2.min.css">
         <style>
             .table-title{
                 background-color: rgb(52,152,219);
@@ -88,25 +84,55 @@
                                                        class="settings" title="Sửa" 
                                                        data-toggle="tooltip"><i class="fas fa-edit" style="margin:0 7px; color:#299BE4;"></i></a>
 
-                                                    <a href="DeleteVaccineServlet?action=delete&id=${vaccine.vaccineID}" 
-                                                       onclick="return confirm('Bạn có chắc chắn?')" class="delete" title="Xóa" 
-                                                       data-toggle="tooltip"><i class="fas fa-trash"style="color: red"></i></a>
+                                                    <a href="DeleteVaccineServlet?action=delete&id=${vaccine.vaccineID}" onclick="confirmDelete(event, this)" class="delete" title="Xóa" data-toggle="tooltip">
+                                                        <i class="fas fa-trash" style="color: red"></i>
+                                                    </a>
 
                                                 </td>
                                             </tr>
                                         </c:forEach>
                                     </tbody>
                                 </table>
-                                
+
                             </div>
-                            
+
                         </div>
                     </div>
                 </div>
             </div>
 
         </div>
-            <script src="assets/vendor/aos/aos.js"></script>
+        <script>
+            $(document).ready(function () {
+                $('[data-toggle="tooltip"]').tooltip();
+            });
+        </script>
+        <script>
+
+
+            function confirmDelete(event, element) {
+                event.preventDefault(); // Ngăn chặn hành động mặc định của liên kết
+                Swal.fire({
+                    title: 'Bạn có chắc chắn?',
+                    
+                    icon: 'warning',
+                    showCancelButton: true,
+                    confirmButtonColor: '#3085d6',
+                    cancelButtonColor: '#d33',
+                    confirmButtonText: 'Vâng, xóa nó!'
+                }).then((result) => {
+                    if (result.isConfirmed) {
+                        Swal.fire(
+                                'Đã xóa vaccine thành công',
+                                ).then(() => {
+                            window.location.href = element.href; // Điều hướng đến liên kết sau khi xác nhận
+                        });
+                    }
+                });
+            }
+        </script>
+        <script src="https://cdn.jsdelivr.net/npm/sweetalert2@11/dist/sweetalert2.min.js"></script>
+        <script src="assets/vendor/aos/aos.js"></script>
         <script src="./assets/libs/jquery/dist/jquery.min.js"></script>
         <script src="./assets/libs/bootstrap/dist/js/bootstrap.bundle.min.js"></script>
         <script src="./assets/js/sidebarmenu.js"></script>
